feat: add callback for setupTask
