+++ benchmark-runner/profile/index.ts
@@ -1,0 +1,12 @@
+import { getWorkspaces, readTaskConfig } from "../utils";
+
+async function run() {
+  const workspaces = getWorkspaces();
+  for (const workspace of workspaces) {
+    const task = readTaskConfig(workspace.path);
+    console.log(task)
+  }
+  
+}
+
+run();
